ivc

Module ivc

Modules

columns

IVC circuit layout - Top level documentation outdated

constraints

helpers

interpreter

lookups

Constants

LIMB_BITSIZE_XLARGE

The biggest packing variant for foreign field. Used for hashing. 150-bit limbs.

N_ADDITIONAL_WIT_COL_QUAD

Number of additional columns that a reduction to degree 2 will require. A regression test is available in the tests directory, under the name test_regression_additional_columns_reduction_to_degree_2

N_ALPHAS

Number of alphas needed for the IVC circuit, equal is the number of all the constraints per row.

N_CONSTRAINTS

Number of constraints used by the IVC circuit.

N_LIMBS_XLARGE

The biggest packing format, 2 limbs.