Enum ivc::ivc::columns::IVCColumn

pub enum IVCColumn {

    FoldIteration,
    BlockSel(usize),
    Block1Input(usize),
    Block1InputRepacked75(usize),
    Block1InputRepacked150(usize),
    Block2Hash(PoseidonColumn<IVC_POSEIDON_STATE_SIZE, IVC_POSEIDON_NB_FULL_ROUND, IVC_POSEIDON_NB_PARTIAL_ROUND>),
    Block3ConstPhi,
    Block3ConstR,
    Block3PhiPow,
    Block3PhiPowR,
    Block3PhiPowR2,
    Block3PhiPowR3,
    Block3PhiPowLimbs(usize),
    Block3PhiPowRLimbs(usize),
    Block3PhiPowR2Limbs(usize),
    Block3PhiPowR3Limbs(usize),
    Block4Input1(usize),
    Block4Coeff,
    Block4Input2AccessTime,
    Block4Input2(usize),
    Block4ECAddInter(FECColumnInter),
    Block4OutputRaw(FECColumnOutput),
    Block4OutputAccessTime,
    Block4OutputRepacked(usize),
    Block5ConstHr,
    Block5ConstR,
    Block5ChalLeft,
    Block5ChalRight,
    Block5ChalOutput,
    Block6ConstR,
    Block6ULeft,
    Block6UOutput,

}

Variants

FoldIteration

A single column containing the folding iteration number, starting with 0 for the base case, and non-zero positive for inductive case.

BlockSel(usize)

Selector for blocks. Inner usize is ∈ [0,#blocks). The selectors are fixed values for each instance of the relation. The different blocks are described below or at the top-level of this file.

Block1Input(usize)

2*17 15-bit limbs (two base field points)

Block1InputRepacked75(usize)

2*4 75-bit limbs

Block1InputRepacked150(usize)

2*2 150-bit limbs

Block2Hash(PoseidonColumn<IVC_POSEIDON_STATE_SIZE, IVC_POSEIDON_NB_FULL_ROUND, IVC_POSEIDON_NB_PARTIAL_ROUND>)

1 hash per row

Block3ConstPhi

Constant phi

Block3ConstR

Constant r

Block3PhiPow

Scalar coeff #1, powers of Phi, phi^i

Block3PhiPowR

Scalar coeff #2, r * phi^i

Block3PhiPowR2

Scalar coeff #2, r^2 * phi^i

Block3PhiPowR3

Scalar coeff #2, r^3 * phi^i

Block3PhiPowLimbs(usize)

17 15-bit limbs

Block3PhiPowRLimbs(usize)

17 15-bit limbs

Block3PhiPowR2Limbs(usize)

17 15-bit limbs

Block3PhiPowR3Limbs(usize)

17 15-bit limbs

Block4Input1(usize)

2*4 75-bit limbs

Block4Coeff

Coeffifient which is indexing a bucket. Used for both lookups in this row.

Block4Input2AccessTime

RAM lookup access time for input 1.

Block4Input2(usize)

2*4 75-bit limbs

Block4ECAddInter(FECColumnInter)

EC ADD intermediate wires

Block4OutputRaw(FECColumnOutput)

2*17 15-bit limbs

Block4OutputAccessTime

RAM lookup access time for output.

Block4OutputRepacked(usize)

2*4 75-bit limbs

Block5ConstHr

Constant h_r

Block5ConstR

Constant r

Block5ChalLeft

α_{L,i}

Block5ChalRight

α_R^i, where α_R = h_R

Block5ChalOutput

α_{O,i} = α_{L,i} + r·α_R^i

Block6ConstR

Constant r

Block6ULeft

u_L

Block6UOutput

u_O = u_L + r