Module ivc::ivc::interpreter

Structs

ScalarLimbs

Contains vectors of scalars in small limb representations. Generic consts don’t allow +1, so vectors not arrays. N is N_COL_TOTAL.

Constants

LIMB_BITSIZE_XLARGE

The biggest packing variant for foreign field. Used for hashing. 150-bit limbs.

N_LIMBS_XLARGE

The biggest packing format, 2 limbs.

Functions

build_selectors

Builds selectors for the IVC circuit. The round constants for Poseidon are not added in this function, and must be done separately. The size of the array is the total number of public values required for the IVC. Therefore, it includes the potential round constants required by the hash function.

combine_large_to_full_field

Helper. Combines large limbs into one element. Computation is over the field.

combine_large_to_xlarge

Helper. Combines small limbs into big limbs.

combine_small_to_full

Helper. Combines 17x15bit limbs into 1 native field element.

constrain_challenges

constrain_ecadds

constrain_inputs

Constraints for the inputs block.

constrain_ivc

This function generates constraints for the whole IVC circuit.

constrain_scalars

constrain_selectors

This function generates constraints for the whole IVC circuit.

constrain_u

ivc_circuit

Instantiates the IVC circuit for folding. L is relaxed (folded) instance, and R is strict (new) instance that is being relaxed at this step. N_COL_TOTAL is the total number of columnsfor IVC + APP.

process_challenges

process_ecadds

process_hashes

Instantiates the IVC circuit for folding. N is the total number of columns

process_inputs

comms is lefts, rights, and outs. Returns the packed commitments in three different representations.

process_scalars

Processes scalars. Returns a vector of limbs of (powers of) scalars produced.

process_u

write_inputs_row

write_scalars_row