ivc::ivc::interpreter

Function ivc_circuit

pub fn ivc_circuit<F, Ff, Env, PParams, const N_COL_TOTAL: usize, const N_CHALS: usize>(
    env: &mut Env,
    fold_iteration: usize,
    comms_left: Box<[(Ff, Ff); N_COL_TOTAL]>,
    comms_right: Box<[(Ff, Ff); N_COL_TOTAL]>,
    comms_out: Box<[(Ff, Ff); N_COL_TOTAL]>,
    error_terms: [(Ff, Ff); 3],
    t_terms: [(Ff, Ff); 2],
    u_l: F,
    chal_l: Box<[F; N_CHALS]>,
    poseidon_params: &PParams,
    domain_size: usize,
)
where
    F: PrimeField,
    Ff: PrimeField,
    PParams: PoseidonParams<F, IVC_POSEIDON_STATE_SIZE, IVC_POSEIDON_NB_TOTAL_ROUND>,
    Env: DirectWitnessCap<F, IVCColumn> + HybridCopyCap<F, IVCColumn> + LookupCap<F, IVCColumn, IVCLookupTable<Ff>>,

Instantiates the IVC circuit for folding. L is relaxed (folded) instance, and R is strict (new) instance that is being relaxed at this step. N_COL_TOTAL is the total number of columns for IVC + APP. N_CHALS is the number of challenges, which contains also the alphas used to combine constraints, see top level documentation in folding. The number of commitments is the total number, and it is expecting the commitments to also the previous IVC columns