ivc::poseidon_55_0_7_3_7::interpreter

Trait PoseidonParams

pub trait PoseidonParams<F: PrimeField, const STATE_SIZE: usize, const NB_FULL_ROUNDS: usize> {
    // Required methods
    fn constants(&self) -> [[F; STATE_SIZE]; NB_FULL_ROUNDS];
    fn mds(&self) -> [[F; STATE_SIZE]; STATE_SIZE];
}

Represents the parameters of the instance of the Poseidon permutation. Constants are the round constants for each round, and MDS is the matrix used by the linear layer. The type is parametrized by the field, the state size, and the number of full rounds. Note that the parameters are only for instances using full rounds.

Required Methods

fn constants(&self) -> [[F; STATE_SIZE]; NB_FULL_ROUNDS]

fn mds(&self) -> [[F; STATE_SIZE]; STATE_SIZE]