ivc::poseidon_8_56_5_3_2

Module bn254

Poseidon parameters that can be used by crate::poseidon_8_56_5_3_2 over the scalar field of BN254, for a security level of 128 bits.

Structs

PlonkSpongeConstantsIVC

Constants used by the IVC circuit used by the folding scheme This is meant to be only used for the IVC circuit, with the BN254 curve It has not been tested with/for other curves.

PoseidonBN254Parameters

Constants

MAX_DEGREE

The maximum degree of a constraint

NB_COLUMNS

The number of columns in the poseidon permutation

NB_CONSTRAINTS

The number of constraints required by this gadget

NB_FULL_ROUND

Number of full rounds

NB_PARTIAL_ROUND

Number of partial rounds

NB_ROUND_CONSTANTS

Number of round constants

NB_TOTAL_ROUND

Total number of rounds, including partial and full.

STATE_SIZE

The number of field elements in the state

Functions

static_params

Type Aliases

Column