ivc

Module poseidon_params_55_0_7_3

Poseidon parameters for 55 full rounds, 0 partial rounds, sbox 7, a state of 3 elements Poseidon parameters that can be used by crate::poseidon_55_0_7_3_2 and crate::poseidon_55_0_7_3_7 over the scalar field of BN254

Structs

PlonkSpongeConstantsIVC

Constants used by the IVC circuit used by the folding scheme This is meant to be only used for the IVC circuit, with the BN254 curve It has not been tested with/for other curves.

Functions

static_params