ivc::ivc

Module interpreter

Structs

ScalarLimbs

Contains vectors of scalars in small limb representations. Generic consts don’t allow +1, so vectors not arrays. N is N_COL_TOTAL.

Functions

build_fixed_selectors

Builds selectors for the IVC circuit. The round constants for Poseidon are not added in this function, and must be done separately. The size of the array is the total number of public values required for the IVC. Therefore, it includes the potential round constants required by the hash function.

ivc_circuit

Instantiates the IVC circuit for folding. L is relaxed (folded) instance, and R is strict (new) instance that is being relaxed at this step. N_COL_TOTAL is the total number of columns for IVC + APP. N_CHALS is the number of challenges, which contains also the alphas used to combine constraints, see top level documentation in folding. The number of commitments is the total number, and it is expecting the commitments to also the previous IVC columns

ivc_circuit_base_case

Base case IVC circuit, completely turned off. For the base case, we do set the fold iteration to 0, and we don’t do any computation. As each constraint is multiplied by the fold iteration, this will simulate a “deactivation” of the IVC circuit.

process_challenges

process_ecadds

process_hashes

Instantiates the IVC circuit for folding. N_COL_TOTAL is the total number of columns required by the IVC circuit + the application.

process_inputs

comms is lefts, rights, and outs. Returns the packed commitments in three different representations.

process_scalars

Processes scalars. Returns a vector of limbs of (powers of) scalars produced.

process_u

write_inputs_row

write_scalars_row