Function ivc::ivc::interpreter::process_hashes

pub fn process_hashes<F, Env, PParams, const N_COL_TOTAL: usize, const N_CHALS: usize>(
    env: &mut Env,
    fold_iteration: usize,
    poseidon_params: &PParams,
    comms_xlarge: &[[[F; 4]; N_COL_TOTAL]; 3]
) -> (Env::Variable, Env::Variable, Env::Variable)where
    F: PrimeField,
    PParams: PoseidonParams<F, IVC_POSEIDON_STATE_SIZE, IVC_POSEIDON_NB_TOTAL_ROUND>,
    Env: MultiRowReadCap<F, IVCColumn> + HybridCopyCap<F, IVCColumn>,

Instantiates the IVC circuit for folding. N_COL_TOTAL is the total number of columns required by the IVC circuit + the application.

The input comms_xlarge contains the commitments to the left, right and output instances (hence the outer size of 3). Each nested N_COL_TOTAL contains the 4 limbs of 150 bits encoding the two coordinates of a commitment.

For instance, if there are 2 columns, the parameter comms_xlarge will be 6 elliptic curve points C_L_1, C_R_1, C_O_1, C_L_2, C_R_2, C_O_2. Each elliptic curve point is represented in affine coordinates by two values (x, y) in the base field of the curve. We split each coordinate into 2 chunks of 150 bits.

We have therefore 12 scalar field elements, as follow:

         C_L_1                          C_L_2
(x_l_1_0_150, x_l_1_150_255) | (x_l_2_0_150, x_l_2_150_255) |
(y_l_1_0_150, y_l_1_150_255) | (y_l_2_0_150, y_l_2_150_255) |
         C_R_1                          C_R_2
(x_r_1_0_150, x_r_1_150_255) | (x_r_1_0_150, x_r_2_150_255) |
(y_r_1_0_150, y_r_1_150_255) | (y_r_1_0_150, y_r_2_150_255) |
         C_O_1                          C_O_2
(x_o_1_0_150, x_o_1_150_255) | (x_o_2_0_150, x_o_2_150_255) |
(y_o_1_0_150, y_o_1_150_255) | (y_o_2_0_150, y_o_2_150_255) |

We will absorb in the following order:

   ---- Left commitments  ---
- `(x_l_1_0_150, x_l_1_150_255)`
- `(y_l_1_0_150, y_l_1_150_255)`
- `(x_l_2_0_150, x_l_2_150_255)`
- `(y_l_2_0_150, y_l_2_150_255)`
   ---- Right commitments  ---
- `(x_r_1_0_150, x_r_1_150_255)`
- `(y_r_1_0_150, y_r_1_150_255)`
- `(x_r_2_0_150, x_r_2_150_255)`
- `(y_r_2_0_150, y_r_2_150_255)`
   ---- Output commitments  ---
- `(x_o_1_0_150, x_o_1_150_255)`
- `(y_o_1_0_150, y_o_1_150_255)`
- `(x_o_2_0_150, x_o_2_150_255)`
- `(y_o_2_0_150, y_o_2_150_255)`

For this:

we get the previous state of the sponge (s0, s1, s2). We copy it in the three first columns of the row.

we constrain the two next columns to be equal to (s0 + i1, s1 + i2). We have then the columns 3, 4 and 5 equal to (s2, s0 + i1, s1 + i2). These will be considered as the input of the Poseidon gadget. This function therefore requires 2 + IVC_POSEIDON_NB_COLUMNS columns, without counting the block selector. It also introduces IVC_POSEIDON_NB_CONSTRAINTS + 2 constraints, and therefore requires IVC_POSEIDON_NB_CONSTRAINTS + 2 alphas.