ivc::poseidon_55_0_7_3_7::interpreter

Function apply_permutation

pub fn apply_permutation<F: PrimeField, const STATE_SIZE: usize, const NB_FULL_ROUND: usize, PARAMETERS, Env>(
    env: &mut Env,
    param: &PARAMETERS,
) -> [Env::Variable; STATE_SIZE]
where
    PARAMETERS: PoseidonParams<F, STATE_SIZE, NB_FULL_ROUND>,
    Env: ColAccessCap<F, PoseidonColumn<STATE_SIZE, NB_FULL_ROUND>> + HybridCopyCap<F, PoseidonColumn<STATE_SIZE, NB_FULL_ROUND>>,

Apply the whole permutation of Poseidon to the state. The environment has to be initialized with the input values.